Fix for createComputeHyperdiskPool.test.js
